truchet: Tiles and Random Mosaics Truchet Style Using the 'sf' Framework

Functions to create Truchet tiles, so called after Sébastien Truchet who was the first to describe the patterns obtained by rotating tiles with respect to each other. This form of tiling is described by Smith and Boucher (1987) https://muse.jhu.edu/article/600574.

See also

Author

Maintainer: Antonio Páez paezha@mcmaster.ca (ORCID)

Authors:

Antonio Páez paezha@mcmaster.ca (ORCID)